requirements.txt
# Web framework
fastapi>=0.128.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0

# Database
sqlalchemy>=2.0.0

# Browser automation
playwright>=1.40.0

# LLM clients
anthropic>=0.34.0
openai>=1.40.0

# Resume parsing
pypdf>=4.0.0
python-docx>=1.1.0

# PDF generation
weasyprint>=62.0

# Configuration
pydantic-settings>=2.4.0
python-dotenv>=1.0.0

# Security
cryptography>=42.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
